requirement.txt
langchain
langchain_community
langchain_google_genai
google-generativeai
streamlit
dotenv
google-api-python-client
pypdf
chromadb